FOP-2265: Enable xlint cast; patch submitted by Simon Steiner
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1502902 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/fonts/FontUtil.java
@@ -40,7 +40,7 @@ public final class FontUtil {
         int weight = 400;
         try {
             weight = Integer.parseInt(text);
-            weight = ((int)weight / 100) * 100;
+            weight = (weight / 100) * 100;
             weight = Math.max(weight, 100);
             weight = Math.min(weight, 900);
         } catch (NumberFormatException nfe) {
